+++ CHANGELOG.md
@@ -24,6 +24,7 @@ Usage:
 * Fix #705: JIB assembly works on Windows
 * Fix #714: feat: Helm support for Golang expressions
 * Port fabric8io/docker-maven-plugin#1318: Update ECR autorization token URL
+* Fix #710: Support DockerImage as output for Openshift builds
 
 ### 1.3.0
 * Fix #497: Assembly descriptor removed but still in documentation